Photon should be running.
I merged the old "player" object with the "playerManager" object from the tutorial. it is now just "PlayerManager". I did the same for the game manager. So currently you start at the launcher scene, connect, join a room, and once the room is joined you load the "TestLevel" scene. once that scene is loaded, if there is no "LocalPlayer", the gameManager instantiates one, and makes the "localPlayer" that instantiated object. then once that player is insantiated, it instantiates a camera prefab, and sets itself as the follow target. I need to do the same thing for the UI, but i havent yet. Im going to build a version of this and see if it works.