Complete overhaul of movement system. Modified netwokring Procedure to interface with new movement system.
I threw the previous movement system into a "deprecated movement" folder in the scripts folder. I put the scripts in a "deprecated" namespace as well so it wouldnt interfere with the new/current scripts. I removed all references to "movementController2D" objects, and replaced all refences to the "movementController" script with a new movementcontroller script that had the same functions, but instead called them on the player's rigidbody2D component. i also just put a static camera on the test scene since the previous one relied on a cc2d component. I'll add tracking to it again later. the PhotonView on the player is also now only tracking the position of the player. These changes have resulted in an incredible improvement to the position synchronization of players. next I should add lag compensation (from photon's documentation) to compensate for the velocity of the rigidbody.