Ravioli ravioli I have found the magic formuoli.
After lots of building and testing, I've found a good formula to sync the position states of players. The lag compensation View sends and recieves a networkedPosition/Velocity Vector, and store that in the instance. On the same instance in fixedUpdate, it interpolates the rigidbody's position with the networked[recieved] position, and sets the rigidbody's velocity to the networked velocity.