Got the Pre game scene working as well as custom room names
added a "room Name" input field to the launcher. Launcher now joins the room with that name, or creates one if there isn't one. Also wrote a script that works in that scene to show a list of players that are in that room, that is updated and reprinted eveytime a player joins or leaves a room.