Added a new CameraFollow Script
Every level will have the camera pre fab already inside it. this will be the client player's camera. When the client's player is instantiated, on the awake callback, it sets the camera's localPlayerGameObject to the local Player. The camera is checking every frame if thte localPlayerGameObject is null, and if not, smoothDamp toward it.